http2: fix integer overflow in unconsumed_bytes_

+++ source/common/http/http2/codec_impl.h
@@ -426,7 +426,7 @@ protected:
     const StreamInfo::BytesMeterSharedPtr& bytesMeter() override { return bytes_meter_; }
     ConnectionImpl& parent_;
     int32_t stream_id_{-1};
-    uint32_t unconsumed_bytes_{0};
+    uint64_t unconsumed_bytes_{0};
     uint32_t read_disable_count_{0};
     StreamInfo::BytesMeterSharedPtr bytes_meter_{std::make_shared<StreamInfo::BytesMeter>()};
